requirements.txt
yt-dlp==2024.4.9
transformers==4.41.0
# torch==2.3.0
openai-whisper==20231117
pyyaml==6.0.1
tiktoken==0.7.0
accelerate==0.30.1
ollama==0.2.0
google-api-python-client==2.129.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
Flask==3.0.3